+++ app/controllers/api/json/visualizations_controller.rb
@@ -21,7 +21,6 @@ class Api::Json::VisualizationsController < Api::ApplicationController
                                                          :is_liked, :remove_like, :index]
   before_filter :optional_api_authorization, only: [:likes_count, :likes_list, :add_like, :is_liked, :remove_like,
                                                     :index]
-  before_filter :link_ghost_tables, only: [:index, :show]
   before_filter :table_and_schema_from_params, only: [:show, :update, :destroy, :stats, :vizjson1, :vizjson2,
                                                       :notify_watching, :list_watching, :likes_count, :likes_list,
                                                       :add_like, :is_liked, :remove_like, :set_next_id]
@@ -118,7 +117,6 @@ class Api::Json::VisualizationsController < Api::ApplicationController
       next_vis.set_prev_list_item!(vis)
     end
 
-    current_user.update_visualization_metrics
     render_jsonp(vis)
   rescue CartoDB::InvalidMember
     render_jsonp({ errors: vis.full_errors }, 400)
@@ -189,7 +187,6 @@ class Api::Json::VisualizationsController < Api::ApplicationController
     vis = Visualization::Member.new(id: @table_id).fetch
     return(head 403) unless vis.is_owner?(current_user)
     vis.delete
-    current_user.update_visualization_metrics
     return head 204
   rescue KeyError
     head(404)
@@ -469,6 +466,10 @@ class Api::Json::VisualizationsController < Api::ApplicationController
     end
   end
 
+  def prepare_params_for_total_count(params)
+      params[:type] == Visualization::Member::TYPE_REMOTE ? params.merge({type: 'table'}) : params
+  end
+
   def index_not_logged_in
     public_visualizations = []
     total_liked_entries = 0
@@ -480,7 +481,7 @@ class Api::Json::VisualizationsController < Api::ApplicationController
       filtered_params = params.dup.merge(scope_for(user))
       filtered_params[:unauthenticated] = true
 
-      total_user_entries = Visualization::Collection.new.count_total(filtered_params)
+      total_user_entries = Visualization::Collection.new.count_total(prepare_params_for_total_count(filtered_params))
 
       collection = Visualization::Collection.new.fetch(filtered_params)
       public_visualizations  = collection.map { |vis|
@@ -515,7 +516,7 @@ class Api::Json::VisualizationsController < Api::ApplicationController
 
     collection = Visualization::Collection.new.fetch(filters)
 
-    total_user_entries = Visualization::Collection.new.count_total(filters)
+    total_user_entries = Visualization::Collection.new.count_total(prepare_params_for_total_count(filters))
 
     table_data = collection.map { |vis|
       if vis.table.nil?
@@ -550,7 +551,6 @@ class Api::Json::VisualizationsController < Api::ApplicationController
       total_likes:    collection.total_liked_entries,
       total_shared:   collection.total_shared_entries
     }
-    current_user.update_visualization_metrics
     render_jsonp(response)
   end