+++ spec/models/visualization/member_spec.rb
@@ -658,6 +658,26 @@ describe Visualization::Member do
                                          }))
       table_member.store
     }.to raise_exception CartoDB::InvalidMember
+
+    child_member = Visualization::Member.new(
+      random_attributes_for_vis_member({
+                                          user_id: @user_mock.id,
+                                          type:       Visualization::Member::TYPE_SLIDE,
+                                          parent_id:  UUIDTools::UUID.timestamp_create.to_s
+                                        }))
+    expect {
+      child_member.store
+    }.to raise_exception CartoDB::InvalidMember
+
+    member_2 = Visualization::Member.new(
+                                         random_attributes_for_vis_member({ user_id: @user_mock.id,
+                                         type: Visualization::Member::TYPE_CANONICAL }))
+    member_2.store.fetch
+    child_member.parent_id = member_2.id
+    expect {
+      child_member.store
+    }.to raise_exception InvalidMember
+
   end
 
   it 'tests transition_options field jsonification' do