+++ NEWS.md
@@ -6,7 +6,21 @@
     - New Config entry: 'subdomains_allowed' (see config/app_config.yml.sample for further details)
 
 
-3.8.0 (2015-01-xx)
+3.8.1 (2015-01-xx)
+------------------
+* Added config for basemaps [#1954], see
+  [documentation](https://github.com/CartoDB/cartodb/wiki/How-to-configure-basemaps-in-CartoDB)
+* Added oEmbed support for visualizations [#1965](https://github.com/CartoDB/cartodb/issues/1965)
+* [content guessing] Prioritize ip over country guessing [#2089](https://github.com/CartoDB/cartodb/issues/2089)
+* Added new import error code (6667) for detecting and reporting statement timeouts.
+* Fixes creation of a visualization from a table [#2145](https://github.com/CartoDB/cartodb/issues/2145)
+* Changes the way geometry types are loaded client side (performance), see PR [#2189](https://github.com/CartoDB/cartodb/pull/218)
+* Cache the geometry types in table model server side (performance), see PR [#2165](https://github.com/CartoDB/cartodb/pull/2165)
+* Added PlatformLimits service. For now, only includes an importer maximum file size limit.
+* Fixed incompatible data for wizards [#1942](https://github.com/CartoDB/cartodb/issues/1942)
+* Added map previews in the delete table dialog
+
+3.8.0 (2015-01-30)
 ------------------
 * Mailchimp user lists importer dataset.
   New Config entry-set: ['oauth']['mailchimp'] (see config/app_config.yml.sample for further details)
@@ -20,6 +34,8 @@
 * Added Mailchimp 'members count' info [#1701](https://github.com/CartoDB/cartodb/issues/1701)
 * Added slides support for visualization
 * Added guessing of types for geojson (boolean and dates) via ogr2ogr2 [#1036](https://github.com/CartoDB/cartodb/issues/1036). Update `ogr2ogr2-static-bin` package for this to work
+* Fixed param `quoted_fields_guessing` on imports and syncs [#1966](https://github.com/CartoDB/cartodb/issues/1966)
+* Background importer for new dashboard.
 
 3.7.1 (2014-12-30)
 ------------------